requirements.txt
black==21.9b0
click==8.0.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
PyYAML==6.0
regex==2021.10.21
SCons==4.2.0
tomli==1.2.1
typing-extensions==3.10.0.2
pyusb
